requirements.txt
fastapi==0.111.0
aiosqlite==0.20.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.37
alembic==1.13.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
websockets==12.0
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
google-generativeai>=0.8.0
